aperturectl control-points list → aperturectl flow-control control-points
