+++ cmd/aperturectl/cmd/root.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/fluxninja/aperture/cmd/aperturectl/cmd/apply"
 	"github.com/fluxninja/aperture/cmd/aperturectl/cmd/blueprints"
-	"github.com/fluxninja/aperture/cmd/aperturectl/cmd/controlpoints"
+	"github.com/fluxninja/aperture/cmd/aperturectl/cmd/flowcontrol"
 	"github.com/fluxninja/aperture/cmd/aperturectl/cmd/installation"
 	"github.com/fluxninja/aperture/pkg/info"
 	"github.com/fluxninja/aperture/pkg/log"
@@ -22,7 +22,7 @@ func init() {
 	RootCmd.AddCommand(apply.ApplyCmd)
 	RootCmd.AddCommand(installation.InstallCmd)
 	RootCmd.AddCommand(installation.UnInstallCmd)
-	RootCmd.AddCommand(controlpoints.ControlPointsCmd)
+	RootCmd.AddCommand(flowcontrol.FlowControlCmd)
 
 	RootCmd.InitDefaultCompletionCmd()
 	RootCmd.SilenceUsage = true
